Upgrade Korge library to v.5.0.6

+++ src/commonMain/kotlin/org/andstatus/game2048/view/Help.kt
@@ -4,7 +4,7 @@ import korlibs.korge.view.position
 
 fun ViewData.showHelp(): MyWindow = myWindow("help_title") {
     wrappableText(
-        stringResources.text("help"), winWidth - 2 * cellMargin, defaultTextSize,
+        stringResources.text("help"), winWidth - 2 * cellMargin, defaultTextSize.toDouble(),
         gameColors.labelText, font, Gravity.LEFT
     ) {
         position(winLeft + cellMargin, winTop + buttonSize + buttonMargin + cellMargin)
